requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pycryptodome
python-multipart
aiofiles
